Expose events for event status testing
